chore : file 디렉토리 정리
