+++ src/List/FriendList/FriendListDesign.tsx
@@ -1,0 +1,16 @@
+import {FriendIntro, FriendName, FriendProfile,IFriendItem } from "../BasicListDesign";
+
+function FriendLists({friend}:IFriendItem){
+
+    return(
+    <>
+    <FriendProfile src = {friend.profile} loading="lazy"/>
+    <FriendName>{friend.name}</FriendName>
+    <FriendIntro introLength ={friend.intro.length}>{friend.intro}</FriendIntro>
+    </>
+
+    );
+
+}
+
+export default FriendLists;
