+++ src/List/ResetUser.tsx
@@ -1,0 +1,10 @@
+import { useResetRecoilState } from "recoil";
+import { userState } from "../recoil";
+
+function ResetUser(){
+    const resetUser = useResetRecoilState(userState);
+    resetUser();
+
+}
+
+export default ResetUser;
