+++ src/Pages/ChatRoom.tsx
@@ -1,13 +1,11 @@
 
-import UserProfile from "./Profile/UserProfile";
-import useMessageList from "./hook/useMessageList";
-import Inputform from "./Input/InputForm";
-import Message from "./Message/Message";
+import UserProfile from "../ChatRoom/Profile/UserProfile";
+import useMessageList from "../hook/useMessageList";
+import Inputform from "../ChatRoom/Input/InputForm";
+import Message from "../ChatRoom/Message/Message";
 import {useParams} from 'react-router-dom';
-import { useState } from "react";
 import { useRecoilState, useResetRecoilState, useSetRecoilState } from "recoil";
-import {messageListState, userState} from "./recoil";
-import chatting from "./data/chat.json";
+import {messageListState, userState} from "../recoil";
 
 function ChatRoom(){
